[리팩터링] 수동로또 생성 로또가 하게끔 위임
